% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fires_summer_uk}
\alias{fires_summer_uk}
\title{Summer Wildfires in the UK (2002-2022)}
\format{
A data frame of class "sf" with 5221 rows and 5 variables:
\describe{
\item{LATITUDE}{Latitude of the fire}
\item{LONGITUDE}{Longitude of the fire}
\item{geometry}{Point coordinates of the fire}
\item{year}{Year}
\item{month}{Month}
}
}
\source{
\url{https://firms.modaps.eosdis.nasa.gov/download/}
}
\usage{
fires_summer_uk
}
\description{
A dataset containing point data of all wildfires that happened in the UK in
the months of June, July, August, andSeptember between 2002 and 2022.
From the MODIS Collection 6.1 of the NASA FIRMS Archive
}
\keyword{datasets}
