% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indic_msoa_eng_wales}
\alias{indic_msoa_eng_wales}
\title{Socioeconomic Indicators for England and Wales at MSOA Level}
\format{
A tibble with 7,264 rows and 26 variables:

\describe{
\item{msoa_code}{\code{character} MSOA code uniquely identifying each area.}
\item{msoa_name}{\code{character} Descriptive name of the MSOA.}
\item{pop_age_15below_normalised}{\code{double} Percentage of the population aged below 15 years old.}
\item{pop_age_65over_normalised}{\code{double} Percentage of the population aged over 65 years old.}
\item{no_qualification_normalised}{\code{double} Percentage of the population without educational qualifications.}
\item{...}{Other socioeconomic indicators such as disability, health status, household composition, housing status, employment, and ethnicity, all normalised by MSOA population.}
}
}
\usage{
indic_msoa_eng_wales
}
\description{
This dataset provides a comprehensive collection of socioeconomic variables
for Middle Layer Super Output Areas (MSOAs) in England and Wales. Derived
primarily from Census data, these indicators have been used for constructing
the Social Vulnerability Index (SoVI).
}
\details{
The indicators within this dataset were selected based on a comprehensive literature review
that identified key factors contributing to social vulnerability, particularly in the context
of wildfire risk. The normalisation process facilitates comparison across different MSOAs
and enhances the dataset's utility in spatial analyses.
}
\references{
The selection of variables is based on the methodology and literature review
conducted in "Spatial Assessment of Wildfire Vulnerability in England and
Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility"
by Hasan Guler.
}
\keyword{datasets}
