% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indic_sdz_ni}
\alias{indic_sdz_ni}
\title{Socioeconomic Indicators for Northern Ireland's Super Data Zones}
\format{
A tibble with 850 rows and 16 variables:

\describe{
\item{sdz21_code}{\code{character} Unique identifier for each Super Data Zone.}
\item{under15_normalised}{\code{double} Proportion of the population under 15 years.}
\item{over65_normalised}{\code{double} Proportion of the population over 65 years.}
\item{no_qual_normalised}{\code{double} Proportion of the population without any formal qualifications.}
\item{disabled_normalised}{\code{double} Proportion of the population with disabilities.}
\item{longterm_condition_normalised}{\code{double} Proportion of the population with long-term health conditions.}
\item{unemployed_normalised}{\code{double} Proportion of the unemployed population.}
\item{skilled_occupation_normalised}{\code{double} Proportion of the population in skilled occupations.}
\item{private_renter_normalised}{\code{double} Proportion of the population living in privately rented accommodations.}
\item{social_renter_normalised}{\code{double} Proportion of the population living in socially rented accommodations.}
\item{no_car_normalised}{\code{double} Proportion of households without access to a car.}
\item{caravan_normalised}{\code{double} Proportion of the population living in caravans or other temporary structures.}
\item{single_person_house_normalised}{\code{double} Proportion of single-person households.}
\item{other_ethnicity_normalised}{\code{double} Proportion of the population belonging to ethnic minorities, excluding the major ethnic groups.}
\item{migrant_inside_normalised}{\code{double} Proportion of the population who migrated from within the UK.}
\item{migrant_outside_normalised}{\code{double} Proportion of the population who migrated from outside the UK.}
}
}
\usage{
indic_sdz_ni
}
\description{
This dataset encapsulates a range of socioeconomic indicators derived from the 2021 Census
for Super Data Zones (SDZs) in Northern Ireland.
}
\details{
The indicators within this dataset were selected based on a comprehensive literature review
that identified key factors contributing to social vulnerability, particularly in the context
of wildfire risk. The normalisation process facilitates comparison across different areas
and enhances the dataset's utility in spatial analyses.
}
\references{
The selection of variables is based on the methodology and literature review
conducted in "Spatial Assessment of Wildfire Vulnerability in England and
Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility"
by Hasan Guler.

Northern Ireland Statistics and Research Agency (NISRA) - Census 2021 data.
}
\keyword{datasets}
