% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_SoVI.R
\name{map_SoVi}
\alias{map_SoVi}
\title{Map Social Vulnerability for a UK Nation}
\usage{
map_SoVi(SoVI_df, nation)
}
\arguments{
\item{SoVI_df}{Dataset including a 'SoVI_standardised' column}

\item{nation}{Nation to be mapped}
}
\value{
An image (png) with the map of social vulnerability
}
\description{
Map the social vulnerability index, either continuous or binned, for a given
nation of the UK at the MSOA level or equivalent.
}
