% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_worst_decile.R
\name{map_worst_decile_msoa}
\alias{map_worst_decile_msoa}
\title{Map Worst Decile MSOAs in the UK}
\usage{
map_worst_decile_msoa(df, nation = "All")
}
\arguments{
\item{df}{Dataset including a 'is_worst_deciles' column}

\item{nation}{(default is "All"). Nation to be mapped: one of "All", "England", "Wales", "Scotland", or "Northern Ireland"}
}
\value{
An image (png) with the map (not yet)
}
\description{
Maps a binary variable at the MSOA level or equivalent for the UK or a given
nation. In this function, the theme is designed for the mapping of a 'worst
quintile' binary variable for the wildfire risk and social vulnerability
index.
}
