% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sovi_scotland}
\alias{sovi_scotland}
\title{Social Vulnerability Index (SoVI) for Scotland's Intermediate Zones}
\format{
A tibble with 1,279 rows and 4 columns:

\describe{
\item{iz11_code}{\code{character} Intermediate Zone code.}
\item{iz11_name}{\code{character} Intermediate Zone name.}
\item{SoVI}{\code{double} Social Vulnerability Index score.}
\item{SoVI_standardised}{\code{double} Standardised Social Vulnerability Index score.}
}
}
\usage{
sovi_scotland
}
\description{
This dataset presents the Social Vulnerability Index (SoVI) for Intermediate Zones (IZs) in Scotland;
The SoVI is a composite measure derived from various socio-economic and demographic variables,
providing insights into relative vulnerability across the UK.
}
\details{
The SoVI is constructed using Principal Component Analysis (PCA) on a set of socio-economic and demographic
variables sourced from the Census, as detailed in the referenced study. This index provides insights into the
relative vulnerability of communities to social and environmental hazards, with higher scores indicating
greater vulnerability.
}
\references{
The methodology for constructing the SoVI is detailed in "Spatial Assessment of Wildfire Vulnerability
in England and Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility," by Hasan Guler.
}
\keyword{datasets}
