% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sovi_wales}
\alias{sovi_wales}
\title{Social Vulnerability Index (SoVI) for Wales' MSOAs}
\format{
A tibble with 408 rows and 4 columns:

\describe{
\item{msoa21_code}{\code{character} MSOA code.}
\item{msoa21_name}{\code{character} MSOA name.}
\item{SoVI}{\code{double} Social Vulnerability Index score.}
\item{SoVI_standardised}{\code{double} Standardised SoVI score.}
}
}
\usage{
sovi_wales
}
\description{
This dataset indicates the Social Vulnerability Index (SoVI) for Middle Layer Super Output Areas (MSOAs) in Wales.
The SoVI is a composite measure derived from various socio-economic and demographic variables,
providing insights into relative vulnerability across the UK.
}
\details{
The SoVI is constructed using Principal Component Analysis (PCA) on a set of socio-economic and demographic
variables sourced from the Census, as detailed in the referenced study. This index provides insights into the
relative vulnerability of communities to social and environmental hazards, with higher scores indicating
greater vulnerability.
}
\references{
The methodology for constructing the SoVI is detailed in the study "Spatial Assessment of Wildfire Vulnerability
in England and Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility," by Hasan Guler.
}
\keyword{datasets}
