% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spring_independent_var_stack}
\alias{spring_independent_var_stack}
\title{Predictors of Spring Wildfires in the UK}
\format{
A RasterStack object with the following layers:

\describe{
\item{Slope}{Raster layer representing the slope of the terrain.}
\item{Aspect}{Raster layer representing the aspect of the terrain.}
\item{Average.Temperature}{Raster layer representing the average temperature during spring.}
\item{Precipitation}{Raster layer representing precipitation during spring.}
\item{Wind.Speed}{Raster layer representing average wind speed during spring.}
\item{Proximity.to.Major.Roads}{Raster layer representing proximity to major roads.}
\item{Population.Counts}{Raster layer representing population counts in the UK.}
}
}
\usage{
spring_independent_var_stack
}
\description{
This RasterStack object contains a collection of raster layers representing
various environmental predictors related to spring wildfires in the UK.
}
\details{
The RasterStack has the following properties:
\itemize{
\item \code{class}: RasterStack
\item \code{dimensions}: 263 rows, 250 columns, 65750 cells, 7 layers
\item \code{resolution}: 0.04166667 x 0.04166667 (x, y)
\item \code{extent}: -8.666667, 1.75, 49.875, 60.83333 (xmin, xmax, ymin, ymax)
\item \code{crs}: +proj=longlat +datum=WGS84 +no_defs
\item \code{names}: Slope, Aspect, Average.Temperature, Precipitation, Wind.Speed, Proximity.to.Major.Roads, Population.Counts
\item \code{min values}: 0.00000, 0.00000, 0.00000, 33.96667, 2.93600, -0.09200, 0.00000
\item \code{max values}: 0.0930622, 6.2831853, 10.5779605, 193.2136383, 10.3826666, 68.3430023, 441.3430481
}
}
\seealso{
\code{\link[raster]{raster}}, \code{\link[raster]{stack}}
}
\author{
Matteo Larrode
}
\keyword{datasets}
