% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summer_independent_var_stack}
\alias{summer_independent_var_stack}
\title{Predictors of Summer Wildfires in the UK}
\format{
A RasterStack object with the following layers:

\describe{
\item{Slope}{Raster layer representing the slope of the terrain.}
\item{Aspect}{Raster layer representing the aspect of the terrain.}
\item{Average.Temperature}{Raster layer representing the average temperature during summer.}
\item{Precipitation}{Raster layer representing precipitation during summer.}
\item{Wind.Speed}{Raster layer representing average wind speed during summer.}
\item{Proximity.to.Major.Roads}{Raster layer representing proximity to major roads.}
\item{Population.Counts}{Raster layer representing population counts in the UK.}
}
}
\usage{
summer_independent_var_stack
}
\description{
This RasterStack object contains a collection of raster layers representing
various environmental predictors related to summer wildfires in the UK.
}
\details{
The RasterStack has the following properties:
\itemize{
\item \code{class}: RasterStack
\item \code{dimensions}: 263 rows, 250 columns, 65750 cells, 7 layers
\item \code{resolution}: 0.04166667 x 0.04166667 (x, y)
\item \code{extent}: -8.666667, 1.75, 49.875, 60.83333 (xmin, xmax, ymin, ymax)
\item \code{crs}: +proj=longlat +datum=WGS84 +no_defs
\item \code{names}: Slope, Aspect, Average.Temperature, Precipitation, Wind.Speed, Proximity.to.Major.Roads, Population.Counts
\item \code{min values}: 0.000000, 0.000000, 8.392517, 44.140907, 2.506667, -0.092000, 0.000000
\item \code{max values}: 0.0930622, 6.2831853, 18.0772878, 203.4393921, 8.6173331, 68.3430023, 441.3430481
}
}
\seealso{
\code{\link[raster]{raster}}, \code{\link[raster]{stack}}
}
\author{
Matteo Larrode
}
\keyword{datasets}
