% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w_sovi_uk}
\alias{w_sovi_uk}
\title{Combined Summer Wildfire Risk and Social Vulnerability Index (SoVI) for UK MSOAs}
\format{
A tibble with 9,393 rows and 5 columns:

\describe{
\item{msoa21_code}{\code{character} MSOA (or equivalent) code.}
\item{ltla21_code}{\code{character} Local Authority code.}
\item{SoVI_standardised}{\code{double} Standardised Social Vulnerability Index score.}
\item{wildfire_risk_standardised}{\code{double} Standardised wildfire risk score.}
\item{is_worst_deciles}{\code{character} Indicator for MSOAs in the worst deciles (8th, 9th, or 10th) for both SoVI and wildfire risk. 'yes' indicates presence in the worst deciles, 'NA' denotes otherwise.}
}
}
\usage{
w_sovi_uk
}
\description{
This dataset integrates the Social Vulnerability Index (SoVI) and summer wildfire risk predictions
across Middle Layer Super Output Areas (MSOAs) in the UK. It includes a binary indicator identifying
MSOAs within the worst deciles (8th, 9th, or 10th) for both SoVI and wildfire risk.
}
\references{
The approach for combining SoVI and wildfire risk scores and the methodology for determining
the worst deciles are based on principles outlined in "Spatial Assessment of Wildfire Vulnerability
in England and Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility" by Hasan Guler.
}
\keyword{datasets}
