% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wildfire_risk_england}
\alias{wildfire_risk_england}
\title{Wildfire Risk Prediction for England's MSOAs}
\format{
A tibble with 6,856 rows and 5 columns:

\describe{
\item{msoa21_name}{\code{character} MSOA name.}
\item{msoa21_code}{\code{character} MSOA code.}
\item{wildfire_risk}{\code{double} Predicted wildfire risk score. Higher values signify greater risk.}
\item{ltla21_code}{\code{character} Local Authority code (higher level geography).}
\item{wildfire_risk_standardised}{\code{double} Standardised wildfire risk score.}
}
}
\usage{
wildfire_risk_england
}
\description{
This dataset provides predicted wildfire risk levels for Middle Layer Super Output Areas (MSOAs)
in England, based on a Random Forest model incorporating various environmental and anthropogenic factors.
}
\details{
Wildfire risk predictions are generated using a Random Forest model, considering climatological,
topographical, and land use variables, as detailed in "Spatial Assessment of Wildfire Vulnerability
in England and Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility" by Hasan Guler.
}
\keyword{datasets}
