% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wildfire_risk_summer_ni}
\alias{wildfire_risk_summer_ni}
\title{Summer Wildfire Risk Prediction for Northern Ireland's Super Data Zones}
\format{
A tibble with 850 rows and 5 columns:

\describe{
\item{sdz21_name}{\code{character} SDZ name.}
\item{sdz21_code}{\code{character} SDZ code.}
\item{wildfire_risk_summer}{\code{double} Predicted wildfire risk score. Higher values signify greater risk.}
\item{ltla21_code}{\code{character} Local Authority code (higher level geography).}
\item{wildfire_risk_summer_standardised}{\code{double} Standardised wildfire risk score.}
}
}
\usage{
wildfire_risk_summer_ni
}
\description{
This dataset outlines predicted wildfire risk levels for Super Data Zones (SDZs) in Northern Ireland in the summer,
derived from a Random Forest analysis that integrates environmental and anthropogenic variables.
}
\details{
Wildfire risk predictions are generated using a Random Forest model, considering climatological,
topographical, and land use variables, as detailed in "Spatial Assessment of Wildfire Vulnerability
in England and Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility" by Hasan Guler.
}
\keyword{datasets}
