% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wildfire_risk_summer_scotland}
\alias{wildfire_risk_summer_scotland}
\title{Summer Wildfire Risk Prediction for Scotland's Intermediate Zones}
\format{
A tibble with 1,279 rows and 5 columns:

\describe{
\item{iz11_name}{\code{character} IZ name.}
\item{iz11_code}{\code{character} IZ code.}
\item{wildfire_risk_summer}{\code{double} Predicted wildfire risk score. Higher values signify greater risk.}
\item{ltla21_code}{\code{character} Local Authority code (higher level geography).}
\item{wildfire_risk_summer_standardised}{\code{double} Standardised wildfire risk score.}
}
}
\usage{
wildfire_risk_summer_scotland
}
\description{
Predicted wildfire risk levels for Intermediate Zones (IZs) in Scotland in the summer, utilizing a Random Forest model
that integrates environmental and anthropogenic variables.
}
\details{
Wildfire risk predictions are generated using a Random Forest model, considering climatological,
topographical, and land use variables, as detailed in "Spatial Assessment of Wildfire Vulnerability
in England and Wales: Coupling Social Vulnerability with Predicted Wildfire Susceptibility" by Hasan Guler.
}
\keyword{datasets}
